refactor: consolidate `renderMarkdown` and `highlightCode` into `handleChat`
