--- src/utils/highlightCode.js
@@ -1,6 +1,0 @@
-export function highlightCode(element) {
-    const codeElements = element.querySelectorAll("pre code");
-    codeElements.forEach((codeElement) => {
-        hljs.highlightElement(codeElement);
-    });
-}
